Added icons to menu selection; fixed buffer big.
So we now have icons for each of the pages designed entirely with
Unicode box-drawing characters. These also slide in and out of the
page when selecting where to go next.

Also, I fixed a bug I did not notice; it was probably born when
I decided to integrate Unicode character handling into the code base,
but basically, when a buffer encounters a -x value, it forgets to
increment j2 to shift the starting position on the passed content array.

+++ src/utils/utils.graphics.h
@@ -1,7 +1,7 @@
 /**
  * @ Author: MMMM
  * @ Create Time: 2024-02-07 02:12:46
- * @ Modified time: 2024-03-08 09:47:48
+ * @ Modified time: 2024-03-10 10:51:56
  * @ Description:
  *    
  * A library that implements graphics-related functionality.
@@ -166,9 +166,9 @@ int Graphics_lerp(color color1, color color2, float fAmount) {
   int b1 = (color1 >> 0) % (1 << 8), b2 = (color2 >> 0) % (1 << 8);
 
   return Graphics_RGB(
-    round(Math_lerp(r1 * 1.0, r2 * 1.0, fAmount)),
-    round(Math_lerp(g1 * 1.0, g2 * 1.0, fAmount)),
-    round(Math_lerp(b1 * 1.0, b2 * 1.0, fAmount))
+    (int) round(Math_lerp(r1 * 1.0, r2 * 1.0, fAmount)),
+    (int) round(Math_lerp(g1 * 1.0, g2 * 1.0, fAmount)),
+    (int) round(Math_lerp(b1 * 1.0, b2 * 1.0, fAmount))
   );
 }
